+++ schema/sql/spNearby.sql
@@ -885,7 +885,7 @@ CREATE FUNCTION fGetNearbySpecObjXYZ (@nx float, @ny float, @nz float, @r float)
 -------------------------------------------------------------
 --/T There is no limit on the number of objects returned, but there are about 40 per sq arcmin.
 --/T <p>returned table:  
---/T <li> specObjID bigint,               -- Photo primary object identifier
+--/T <li> specObjID numeric(20,0),               -- Photo primary object identifier
 --/T <li> run int NOT NULL,           -- run that observed this object   
 --/T <li> camcol int NOT NULL,        -- camera column that observed the object
 --/T <li> field int NOT NULL,         -- field that had the object
@@ -905,7 +905,7 @@ CREATE FUNCTION fGetNearbySpecObjXYZ (@nx float, @ny float, @nz float, @r float)
 --/T <br>see also fGetNearbySpecObjEq, fGetNearestSpecObjXYZ, fGetNearbySpecObjAllXYZ
 -------------------------------------------------------------
   RETURNS @proxtab TABLE (
-    specObjID bigint,
+    specObjID numeric(20,0),
     plate int NOT NULL,
     mjd int NOT NULL,
     fiberID int NOT NULL,
@@ -982,7 +982,7 @@ CREATE FUNCTION fGetNearbySpecObjAllXYZ (@nx float, @ny float, @nz float, @r flo
 --/T <br>see also fGetNearbySpecObjEq, fGetNearestSpecObjXYZ, fGetNearestSpecObjXYZ
 -------------------------------------------------------------
   RETURNS @proxtab TABLE (
-    specObjID bigint,
+    specObjID numeric(20,0),
     plate int NOT NULL,
     mjd int NOT NULL,
     fiberID int NOT NULL,
@@ -1040,7 +1040,7 @@ CREATE FUNCTION fGetNearestSpecObjXYZ (@nx float, @ny float, @nz float, @r float
 --/H Returns nearest scienceprimary specobj within @r arcminutes of an xyz point (@nx,@ny, @nz).
 -------------------------------------------------------------
 --/T <p> returned table:  
---/T <li>  specObjID bigint,		-- unique spectrum ID
+--/T <li>  specObjID numeric(20,0),		-- unique spectrum ID
 --/T <li>  plate int NOT NULL,		-- spectroscopic plate number
 --/T <li>  mjd int NOT NULL,		-- MJD of observation
 --/T <li>  fiberID int NOT NULL,	-- fiber number on plate
@@ -1061,7 +1061,7 @@ CREATE FUNCTION fGetNearestSpecObjXYZ (@nx float, @ny float, @nz float, @r float
 --/T  <br>see also fGetNearbySpecObjEq, fGetNearestSpecObjEq, fGetNearbySpecObjXYZ, 
 -------------------------------------------------------------
   RETURNS @proxtab TABLE (
-    specObjID bigint,
+    specObjID numeric(20,0),
     plate int NOT NULL,
     mjd int NOT NULL,
     fiberID int NOT NULL,
@@ -1094,7 +1094,7 @@ CREATE FUNCTION fGetNearestSpecObjAllXYZ (@nx float, @ny float, @nz float, @r fl
 --/H Returns nearest specobj within @r arcminutes of an xyz point (@nx,@ny, @nz).
 -------------------------------------------------------------
 --/T <p> returned table:  
---/T <li>  specObjID bigint,		-- unique spectrum ID
+--/T <li>  specObjID numeric(20,0),		-- unique spectrum ID
 --/T <li>  plate int NOT NULL,		-- spectroscopic plate number
 --/T <li>  mjd int NOT NULL,		-- MJD of observation
 --/T <li>  fiberID int NOT NULL,	-- fiber number on plate
@@ -1116,7 +1116,7 @@ CREATE FUNCTION fGetNearestSpecObjAllXYZ (@nx float, @ny float, @nz float, @r fl
 --/T  <br>see also fGetNearbySpecObjAllEq, fGetNearestSpecObjAllEq, fGetNearbySpecObjAllXYZ, 
 -------------------------------------------------------------
   RETURNS @proxtab TABLE (
-    specObjID bigint,
+    specObjID numeric(20,0),
     plate int NOT NULL,
     mjd int NOT NULL,
     fiberID int NOT NULL,
@@ -1151,7 +1151,7 @@ CREATE FUNCTION fGetNearbySpecObjEq (@ra float, @dec float, @r float)
 -------------------------------------------------------------
 --/T There is no limit on the number of objects returned.
 --/T <p>returned table:  
---/T <li>  specObjID bigint,		-- unique spectrum ID
+--/T <li>  specObjID numeric(20,0),		-- unique spectrum ID
 --/T <li>  plate int NOT NULL,		-- spectroscopic plate number
 --/T <li>  mjd int NOT NULL,		-- MJD of observation
 --/T <li>  fiberID int NOT NULL,	-- fiber number on plate
@@ -1171,7 +1171,7 @@ CREATE FUNCTION fGetNearbySpecObjEq (@ra float, @dec float, @r float)
 --/T <br>see also fGetNearbySpecObjEq, fGetNearestSpecObjXYZ, fGetNearestSpecObjXYZ
 -------------------------------------------------------------
   RETURNS @proxtab TABLE (
-    specObjID bigint,
+    specObjID numeric(20,0),
     plate int NOT NULL,
     mjd int NOT NULL,
     fiberID int NOT NULL,
@@ -1209,7 +1209,7 @@ CREATE FUNCTION fGetNearbySpecObjAllEq (@ra float, @dec float, @r float)
 -------------------------------------------------------------
 --/T There is no limit on the number of objects returned.
 --/T <p>returned table:  
---/T <li>  specObjID bigint,		-- unique spectrum ID
+--/T <li>  specObjID numeric(20,0),		-- unique spectrum ID
 --/T <li>  plate int NOT NULL,		-- spectroscopic plate number
 --/T <li>  mjd int NOT NULL,		-- MJD of observation
 --/T <li>  fiberID int NOT NULL,	-- fiber number on plate
@@ -1230,7 +1230,7 @@ CREATE FUNCTION fGetNearbySpecObjAllEq (@ra float, @dec float, @r float)
 --/T <br>see also fGetNearbySpecObjEq, fGetNearestSpecObjXYZ, fGetNearestSpecObjXYZ
 -------------------------------------------------------------
   RETURNS @proxtab TABLE (
-    specObjID bigint,
+    specObjID numeric(20,0),
     plate int NOT NULL,
     mjd int NOT NULL,
     fiberID int NOT NULL,
@@ -1268,7 +1268,7 @@ CREATE FUNCTION fGetNearestSpecObjEq (@ra float, @dec float, @r float)
 --/H Returns nearest scienceprimary specobj within @r arcminutes of an equatorial point (@ra,@dec).
 -------------------------------------------------------------
 --/T <p> returned table:  
---/T <li>  specObjID bigint,		-- unique spectrum ID
+--/T <li>  specObjID numeric(20,0),		-- unique spectrum ID
 --/T <li>  plate int NOT NULL,		-- spectroscopic plate number
 --/T <li>  mjd int NOT NULL,		-- MJD of observation
 --/T <li>  fiberID int NOT NULL,	-- fiber number on plate
@@ -1289,7 +1289,7 @@ CREATE FUNCTION fGetNearestSpecObjEq (@ra float, @dec float, @r float)
 --/T  <br>see also fGetNearbySpecObjEq, fGetNearestSpecObjAllEq, fGetNearbySpecObjXYZ, 
 -------------------------------------------------------------
   RETURNS @proxtab TABLE (
-    specObjID bigint,
+    specObjID numeric(20,0),
     plate int NOT NULL,
     mjd int NOT NULL,
     fiberID int NOT NULL,
@@ -1328,7 +1328,7 @@ CREATE FUNCTION fGetNearestSpecObjAllEq (@ra float, @dec float, @r float)
 --/H Returns nearest specobj within @r arcminutes of an equatorial point (@ra, @dec).
 -------------------------------------------------------------
 --/T <p> returned table:  
---/T <li>  specObjID bigint,		-- unique spectrum ID
+--/T <li>  specObjID numeric(20,0),		-- unique spectrum ID
 --/T <li>  plate int NOT NULL,		-- spectroscopic plate number
 --/T <li>  mjd int NOT NULL,		-- MJD of observation
 --/T <li>  fiberID int NOT NULL,	-- fiber number on plate
@@ -1350,7 +1350,7 @@ CREATE FUNCTION fGetNearestSpecObjAllEq (@ra float, @dec float, @r float)
 --/T  <br>see also fGetNearbySpecObjAllEq, fGetNearestSpecObjEq, fGetNearbySpecObjAllXYZ, 
 -------------------------------------------------------------
   RETURNS @proxtab TABLE (
-    specObjID bigint,
+    specObjID numeric(20,0),
     plate int NOT NULL,
     mjd int NOT NULL,
     fiberID int NOT NULL,
@@ -1942,7 +1942,7 @@ CREATE  PROCEDURE spGetSpecNeighborsRadius
 --/T specObjId. The result of this is then joined with the SpecObj 
 --/T table, to return the attributes of the photometric objects.
 --/T <samp>
---/T <br> create table #x (id int,specObjID bigint)
+--/T <br> create table #x (id int,specObjID numeric(20,0))
 --/T <br> insert into #x EXEC spGetNeighbours 
 --/T </samp>
 -------------------------------------------------------------------
@@ -1973,7 +1973,7 @@ CREATE PROCEDURE spGetSpecNeighborsPrim(@r float)
 --/T The result of this is then joined with the SpecObj table, 
 --/T to return the attributes of the photometric objects.
 --/T <samp>
---/T <br> create table #x (id int,specObjID bigint)
+--/T <br> create table #x (id int,specObjID numeric(20,0))
 --/T <br> insert into #x EXEC spGetSpecNeighborsPrim 2.5
 --/T </samp>
 -------------------------------------------------------------------
@@ -2004,7 +2004,7 @@ CREATE PROCEDURE spGetSpecNeighborsAll(@r float)
 --/T The result of this is then joined with the SpecObjAll table, 
 --/T to return the attributes of the photometric objects.
 --/T <samp>
---/T <br> create table #x (id int,specObjID bigint)
+--/T <br> create table #x (id int,specObjID numeric(20,0))
 --/T <br> insert into #x EXEC spGetSpecNeighborsAll 2.5
 --/T </samp>
 -------------------------------------------------------------------
